switch -> object로 형태변경, 스타일 유틸함수로 분리
